python chatbot/requirements.txt
flask==3.0.0
openai==1.51.2
python-dotenv==1.0.0
gunicorn==21.2.0
